Refacto: externalize the performance test generation into the core-performance package
