+++ packages/core-performance/src/utils/api.js
@@ -1,0 +1,15 @@
+function clean(obj) {
+  return Object.keys(obj).reduce((result, key) => {
+    if (obj[key] !== undefined) result[key] = obj[key];
+    return result;
+  }, {});
+}
+
+function getContentType(headerContentType) {
+  return String(headerContentType.split("/").pop()).split(";").shift();
+}
+
+module.exports = {
+  getContentType,
+  clean
+};
