+++ packages/core-performance/src/utils/process.js
@@ -1,0 +1,3 @@
+module.exports = {
+  cwd: process.cwd()
+};
