+++ packages/core-performance/src/vendors/index.js
@@ -1,0 +1,6 @@
+module.exports = {
+  artillery: require("./artillery")
+  // k6: require("./k6"),
+  // jmeter: require("./jmeter"),
+  // vegeta: require("./vegeta")
+};
