+++ packages/core-performance/src/vendors/k6.js
@@ -1,0 +1,3 @@
+module.exports = function (features, outputFolder) {
+  // @todo
+};
