+++ packages/core-performance/src/vendors/vegeta.js
@@ -1,0 +1,3 @@
+module.exports = function (features, outputFolder) {
+  // @todo
+};
